Added Additional functionalities
Delete request for users will delete users and delete the user from the project they are listed under

Delete request for project will delete the project and remove the project from the user's project field

A Get method to get projects with the badge names only with the enum field.

User patch request will update the user details, name, username, email, password and usertype. Assumes that there will be a method to remove comments, projects and likeProjects later in development